Cleanup and refactoring
- Update listing summary card containers to use `listingPhaseState`
selector data
- Rename selector file to a more sensible one
